+++ app/src/renderer/components/staking/TabMyDelegations.vue
@@ -4,15 +4,8 @@
       | Your Validators
       |
       i.material-icons.info-button(v-tooltip.top="bondInfo") info_outline
-    panel-sort(:sort='sort', :properties="properties")
-    data-empty-search(v-if="yourValidators.length === 0")
-    template(v-else)
-      ol
-        li-validator(
-          v-for='validator in yourValidators'
-          :key='`Your/${validator.id}`'
-          :validator='validator'
-        )
+    table-validators(:validators="yourValidators")
+
     .check-out-message
       | Check out
       |
@@ -25,102 +18,31 @@
         | Your Undelegated Validators
         |
         i.material-icons.info-button(v-tooltip.top="unbondInfo") info_outline
-      ol
-        li-validator(
-          v-for='validator in undelegatedValidators'
-          :key='`Undelegated/${validator.id}`'
-          :validator='validator'
-        )
+      table-validators(:validators="undelegatedValidators")
 </template>
 
 <script>
 import { mapGetters } from "vuex"
-import LiValidator from "staking/LiValidator"
-import { TmDataEmpty } from "@tendermint/ui"
-import DataEmptySearch from "common/TmDataEmptySearch"
-import PanelSort from "staking/PanelSort"
+import TableValidators from "staking/TableValidators"
+
 export default {
   name: `tab-my-delegations`,
-  components: {
-    LiValidator,
-    TmDataEmpty,
-    DataEmptySearch,
-    PanelSort
-  },
+  components: { TableValidators },
   data: () => ({
-    sort: {
-      property: `percent_of_vote`,
-      order: `desc`
-    },
     bondInfo: `Validators you are currently bonded to`,
     unbondInfo: `Your bonded validators in unbonding process`
   }),
   computed: {
-    ...mapGetters([
-      `delegates`,
-      `delegation`,
-      `committedDelegations`,
-      `user`,
-      `bondingDenom`
-    ]),
-    address({ user } = this) {
-      return user.address
-    },
+    ...mapGetters([`delegates`, `delegation`, `committedDelegations`]),
     undelegatedValidators(
       { delegates: { delegates }, delegation: { unbondingDelegations } } = this
     ) {
-      const unbonding = new Set(Object.keys(unbondingDelegations))
-      return delegates.filter(({ id }) => unbonding.has(id))
-    },
-    properties({ bondingDenom } = this) {
-      return [
-        {
-          title: `Moniker`,
-          value: `small_moniker`,
-          tooltip: `The validator's moniker`,
-          class: `name`
-        },
-        {
-          title: `Bonded ${bondingDenom}`,
-          value: `your_votes`,
-          tooltip: `Number of ${bondingDenom} you have delegated to the validator`,
-          class: `your-votes`
-        },
-        {
-          title: `Rewards`,
-          value: `your_rewards`, // TODO: use real rewards
-          tooltip: `Rewards of ${bondingDenom} you have gained from the validator`,
-          class: `your-rewards` // TODO: use real rewards
-        },
-        {
-          title: `Voting Power`,
-          value: `percent_of_vote`,
-          tooltip: `Percentage of ${bondingDenom} the validator has on The Cosmos Hub`,
-          class: `percent_of_vote`
-        },
-        {
-          title: `Uptime`,
-          value: `uptime`,
-          tooltip: `Ratio of blocks signed within the last 10k blocks`,
-          class: `uptime`
-        },
-        {
-          title: `Commission`,
-          value: `commission`,
-          tooltip: `The validator's commission`,
-          class: `commission`
-        },
-        {
-          title: `Slashes`,
-          value: `slashes`, // TODO: use real slashes
-          tooltip: `The validator's slashes`,
-          class: `slashes`
-        }
-      ]
+      const unbonding = Object.keys(unbondingDelegations)
+      return delegates.filter(({ owner }) => unbonding.includes(owner))
     },
     yourValidators({ committedDelegations, delegates: { delegates } } = this) {
-      const committed = new Set(Object.keys(committedDelegations))
-      return delegates.filter(({ id }) => committed.has(id))
+      const committed = Object.keys(committedDelegations)
+      return delegates.filter(({ owner }) => committed.includes(owner))
     }
   }
 }
@@ -129,10 +51,10 @@ export default {
 @require '~variables'
 
 .tab-header
-  margin 1rem 1rem 0 2rem
-  font-size 14px
   color var(--dim)
+  font-size 14px
   font-weight 500
+  margin 1rem 1rem 0 2rem
 
 .info-button
   color var(--link)
